+++ src/models/bradley-terry-full.ts
@@ -1,9 +1,10 @@
-import util, { score } from '../util'
+import util, { score, updatePlayer } from '../util'
 import constants from '../constants'
 import { Rating, Options, Model } from '../types'
 
 const model: Model = (game: Rating[][], options: Options = {}) => {
   const { TWOBETASQ, KAPPA } = constants(options)
+  const { weight } = options
   const { teamRating, gamma } = util(options)
   const teamRatings = teamRating(game)
 
@@ -25,13 +26,7 @@ const model: Model = (game: Rating[][], options: Options = {}) => {
         { omega: 0, delta: 0 }
       )
 
-    return iTeam.map(({ mu, sigma }) => {
-      const sigmaSq = sigma * sigma
-      return {
-        mu: mu + (sigmaSq / iSigmaSq) * iOmega,
-        sigma: sigma * Math.sqrt(Math.max(1 - (sigmaSq / iSigmaSq) * iDelta, KAPPA)),
-      }
-    })
+    return iTeam.map((player, j) => updatePlayer(player, iOmega, iDelta, iSigmaSq, weight?.[i]?.[j] ?? 1, KAPPA))
   })
 }